@@ -1,37 +1,37 @@
-import { containsUnsafe } from "../../src/util";
+import { containsUnsafe } from "../../src/util/index";
 
 describe("Expect package (sanitization):", () => {
   it("blocks &", () => {
     const unsafe = containsUnsafe({
-      value: "test.x.y.&.fest.z"
+      value: "test.x.y.&.fest.z",
     });
     expect(unsafe).toBe(true);
   });
 
   it("blocks strings with surrogate pairs correctly", () => {
     const unsafe = containsUnsafe({
-      value: "&日本語&"
+      value: "&日本語&",
     });
     expect(unsafe).toBe(true);
   });
 
   it("blocks < and >", () => {
     const unsafe = containsUnsafe({
-      value: "<div>This is a html element</div>"
+      value: "<div>This is a html element</div>",
     });
     expect(unsafe).toBe(true);
   });
 
   it("blocks \" and '", () => {
     const unsafe = containsUnsafe({
-      value: 'attribute="test"'
+      value: 'attribute="test"',
     });
     expect(unsafe).toBe(true);
   });
 
   it("does not block ( or ) if not in strict mode", () => {
     const unsafe = containsUnsafe({
-      value: "(do not block this)"
+      value: "(do not block this)",
     });
     expect(unsafe).toBe(false);
   });
@@ -39,14 +39,14 @@ describe("Expect package (sanitization):", () => {
   it("blocks ( and ) if in strict mode", () => {
     const unsafe = containsUnsafe({
       value: "(block this)",
-      strict: true
+      strict: true,
     });
     expect(unsafe).toBe(true);
   });
 
   it("does not block { or } if not in strict mode", () => {
     const unsafe = containsUnsafe({
-      value: "{do not block this}"
+      value: "{do not block this}",
     });
     expect(unsafe).toBe(false);
   });
@@ -54,14 +54,14 @@ describe("Expect package (sanitization):", () => {
   it("blocks { and } if in strict mode", () => {
     const unsafe = containsUnsafe({
       value: "{block this}",
-      strict: true
+      strict: true,
     });
     expect(unsafe).toBe(true);
   });
 
   it("does not block [ or ] if not in strict mode", () => {
     const unsafe = containsUnsafe({
-      value: "[do not block this]"
+      value: "[do not block this]",
     });
     expect(unsafe).toBe(false);
   });
@@ -69,14 +69,14 @@ describe("Expect package (sanitization):", () => {
   it("block [ and ] if in strict mode", () => {
     const unsafe = containsUnsafe({
       value: "[block this]",
-      strict: true
+      strict: true,
     });
     expect(unsafe).toBe(true);
   });
 
   it("does not block ! if not in strict mode", () => {
     const unsafe = containsUnsafe({
-      value: "do not block this!"
+      value: "do not block this!",
     });
     expect(unsafe).toBe(false);
   });
@@ -84,14 +84,14 @@ describe("Expect package (sanitization):", () => {
   it("blocks ! if in strict mode", () => {
     const unsafe = containsUnsafe({
       value: "block this!",
-      strict: true
+      strict: true,
     });
     expect(unsafe).toBe(true);
   });
 
   it("does not block @ if not in strict mode", () => {
     const unsafe = containsUnsafe({
-      value: "do not block this@"
+      value: "do not block this@",
     });
     expect(unsafe).toBe(false);
   });
@@ -99,14 +99,14 @@ describe("Expect package (sanitization):", () => {
   it("blocks @ if in strict mode", () => {
     const unsafe = containsUnsafe({
       value: "block this@",
-      strict: true
+      strict: true,
     });
     expect(unsafe).toBe(true);
   });
 
   it("does not block $ if not in strict mode", () => {
     const unsafe = containsUnsafe({
-      value: "do not block this$"
+      value: "do not block this$",
     });
     expect(unsafe).toBe(false);
   });
@@ -114,14 +114,14 @@ describe("Expect package (sanitization):", () => {
   it("blocks $ if in strict mode", () => {
     const unsafe = containsUnsafe({
       value: "block this$",
-      strict: true
+      strict: true,
     });
     expect(unsafe).toBe(true);
   });
 
   it("does not block = if not in strict mode", () => {
     const unsafe = containsUnsafe({
-      value: "do not block this="
+      value: "do not block this=",
     });
     expect(unsafe).toBe(false);
   });
@@ -129,14 +129,14 @@ describe("Expect package (sanitization):", () => {
   it("blocks = if in strict mode", () => {
     const unsafe = containsUnsafe({
       value: "block this=",
-      strict: true
+      strict: true,
     });
     expect(unsafe).toBe(true);
   });
 
   it("does not block + if not in strict mode", () => {
     const unsafe = containsUnsafe({
-      value: "do not block this+"
+      value: "do not block this+",
     });
     expect(unsafe).toBe(false);
   });
@@ -144,7 +144,7 @@ describe("Expect package (sanitization):", () => {
   it("blocks + if in strict mode", () => {
     const unsafe = containsUnsafe({
       value: "block this+",
-      strict: true
+      strict: true,
     });
     expect(unsafe).toBe(true);
   });
@@ -153,7 +153,7 @@ describe("Expect package (sanitization):", () => {
     const unsafe = containsUnsafe({
       value: "test@sanitize.com",
       strict: true,
-      allowed: ["@"]
+      allowed: ["@"],
     });
     expect(unsafe).toBe(false);
   });
@@ -162,7 +162,7 @@ describe("Expect package (sanitization):", () => {
     const unsafe = containsUnsafe({
       value: "Don't block this text",
       strict: true,
-      allowed: ["'"]
+      allowed: ["'"],
     });
     expect(unsafe).toBe(false);
   });
@@ -171,7 +171,7 @@ describe("Expect package (sanitization):", () => {
     const unsafe = containsUnsafe({
       value: "Some basic math like 1 + 2 = 3 should be allowed! :)",
       strict: true,
-      allowed: ["+", "=", "!", ")"]
+      allowed: ["+", "=", "!", ")"],
     });
     expect(unsafe).toBe(false);
   });
@@ -179,7 +179,7 @@ describe("Expect package (sanitization):", () => {
   it("blocks strings with surrogate pairs correctly in strict mode", () => {
     const unsafe = containsUnsafe({
       value: '<p>"Japanese" in japanese is 日本語!</p>',
-      strict: true
+      strict: true,
     });
     expect(unsafe).toBe(true);
   });
@@ -188,7 +188,7 @@ describe("Expect package (sanitization):", () => {
     const unsafe = containsUnsafe({
       value:
         '<div class="test" data-someattribute="hello">Wow, this sure will get safe!</div>',
-      strict: true
+      strict: true,
     });
     expect(unsafe).toBe(true);
   });
